initial reorganization of imports and lazy load
