+++ docs/api-reference/cortex.md
@@ -1,4 +1,4 @@
-::: centimators.keras_cortex
+::: centimators.model_estimators.cortex
     options:
       show_symbol_type_toc: true
       members_order: source
