+++ docs/api-reference/model_estimators.md
@@ -1,5 +1,9 @@
 # Model Estimators
 
-::: centimators.model_estimators
+::: centimators.model_estimators.keras_estimators
+    options:
+      show_symbol_type_toc: true
+
+::: centimators.model_estimators.dspymator
     options:
       show_symbol_type_toc: true
